requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
playwright==1.46.0
PyMuPDF==1.24.9
python-dotenv==1.0.1
supabase==2.4.0
pandas==2.2.2
tenacity==8.2.3
